test: Do not hardcode expected Snippet IDs
They differ based on how many tests are run. It's been a struggle to
figure out how to reset database in between tests.

Resetting DBs in-between tests in the same run is discouraged
anyway since it will result to slow test runs so not a priority
to figure it out.